Update diamond.py
Added numeration to represent the teams, making the code more readable and maintainable. The logic for setting the glow color and enabling it is extracted into a separate function called set_glow(), making the code more organized and easy to understand.
It also added a key 'q' to turn off the script, and a check to detect if the game is running on a 64-bit or 32-bit architecture and adjust the memory addresses accordingly.